refactor: split service
