+++ typings/app/service/index.d.ts
@@ -4,10 +4,12 @@
 import 'egg';
 import ExportTest from '../../../app/service/Test';
 import ExportGithub from '../../../app/service/github';
+import ExportShimo from '../../../app/service/shimo';
 
 declare module 'egg' {
   interface IService {
     test: ExportTest;
     github: ExportGithub;
+    shimo: ExportShimo;
   }
 }
